requirements.txt
Django==5.2.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-csp==3.7
Pillow==10.2.0
requests==2.31.0
django-environ==0.11.2
django-storages==1.14.2
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-debug-toolbar==4.3.0
paypalrestsdk==1.13.1
djangorestframework>=3.14.0
django-filter>=23.5 